+++ Core/Instrument/IHistogram.h
@@ -96,10 +96,10 @@ public:
 
     //! Returns x-axis bin index for given globalbin. For 1D histograms returned value conicide
     //! with globalbin value.
-    int getXaxisIndex(size_t i) const;
+	size_t getXaxisIndex(size_t i) const;
 
     //! Returns y-axis bin index for given globalbin (for 2D histograms).
-    int getYaxisIndex(size_t i) const;
+	size_t getYaxisIndex(size_t i) const;
 
     //! @brief Returns the value on x-axis corresponding to the global bin index.
     //! @param globalbin The global bin index
