+++ Core/Multilayer/MultiLayer.cpp
@@ -165,7 +165,7 @@ double MultiLayer::crossCorrSpectralFun(const kvector_t kvec, size_t j, size_t k
 
 int MultiLayer::indexOfLayer(const Layer* p_layer) const
 {
-    for (size_t i=0; i<numberOfLayers(); ++i)
+    for (int i=0; i<numberOfLayers(); ++i)
         if (p_layer == m_layers[i])
             return i;
     return -1;
